==> Data and Processing Code/Code and Control Files/2a_combine_simulation_results.py <==
#%% ABOUT
'''
This program combines the output files for all scenarios created by the AHLE
simulation model and adds basic adjustments.

UPDATE August 2023: this program has been updated to work with updated AHLE
scenario results. Murdoch University has updated the compartmental model code
for small ruminants (sheep and goats) and the parameters for both small
ruminants and cattle. The relevant changes for this code are:
    - The small ruminant results from the updated compartmental model contain
    all scenarios in a single file instead of separate files.
    - The updated parameters for cattle are for a single year and for the
    national level only.

IMPORTANT: before running this, set Python's working directory to the folder
where this code is stored.
'''
#%% PACKAGES AND FUNCTIONS

import os                # Operating system functions
import inspect
import io
import time
import numpy as np
import pandas as pd
import scipy.stats as sps
import pickle            # To save objects to disk

# To clean up column names in a dataframe
def cleancolnames(INPUT_DF):
   # Comments inside the statement create errors. Putting all comments at the top.
   # Convert to lowercase
   # Strip leading and trailing spaces, then replace spaces with underscore
   # Replace slashes, parenthesis, and brackets with underscore
   # Replace some special characters with underscore
   # Replace other special characters with words
   INPUT_DF.columns = INPUT_DF.columns.str.lower() \
      .str.strip().str.replace(' ' ,'_' ,regex=False) \
      .str.replace('/' ,'_' ,regex=False).str.replace('\\' ,'_' ,regex=False) \
      .str.replace('(' ,'_' ,regex=False).str.replace(')' ,'_' ,regex=False) \
      .str.replace('[' ,'_' ,regex=False).str.replace(']' ,'_' ,regex=False) \
      .str.replace('{' ,'_' ,regex=False).str.replace('}' ,'_' ,regex=False) \
      .str.replace('!' ,'_' ,regex=False).str.replace('?' ,'_' ,regex=False) \
      .str.replace('-' ,'_' ,regex=False).str.replace('+' ,'_' ,regex=False) \
      .str.replace('^' ,'_' ,regex=False).str.replace('*' ,'_' ,regex=False) \
      .str.replace('.' ,'_' ,regex=False).str.replace(',' ,'_' ,regex=False) \
      .str.replace('|' ,'_' ,regex=False).str.replace('#' ,'_' ,regex=False) \
      .str.replace('>' ,'_gt_' ,regex=False) \
      .str.replace('<' ,'_lt_' ,regex=False) \
      .str.replace('=' ,'_eq_' ,regex=False) \
      .str.replace('@' ,'_at_' ,regex=False) \
      .str.replace('$' ,'_dol_' ,regex=False) \
      .str.replace('%' ,'_pct_' ,regex=False) \
      .str.replace('&' ,'_and_' ,regex=False)
   return None

# To print df.info() with header for readability, and optionally write data info to text file
def datainfo(
      INPUT_DF
      ,MAX_COLS=100
      ,OUTFOLDER=None     # String (opt): folder to output {dataname}_info.txt. If None, no file will be created.
   ):
   funcname = inspect.currentframe().f_code.co_name
   dataname = [x for x in globals() if globals()[x] is INPUT_DF][0]
   rowcount = INPUT_DF.shape[0]
   colcount = INPUT_DF.shape[1]
   idxcols = str(list(INPUT_DF.index.names))
   header = f"Data name: {dataname :>26s}\nRows:      {rowcount :>26,}\nColumns:   {colcount :>26,}\nIndex:     {idxcols :>26s}\n"
   divider = ('-'*26) + ('-'*11) + '\n'
   bigdivider = ('='*26) + ('='*11) + '\n'
   print(bigdivider + header + divider)
   INPUT_DF.info(max_cols=MAX_COLS)
   print(divider + f"End:       {dataname:>26s}\n" + bigdivider)

   if OUTFOLDER:     # If something has been passed to OUTFOLDER parameter
      filename = f"{dataname}_info"
      print(f"\n<{funcname}> Creating file {OUTFOLDER}\{filename}.txt")
      datetimestamp = 'Created on ' + time.strftime('%Y-%m-%d %X', time.gmtime()) + ' UTC' + '\n'
      buffer = io.StringIO()
      INPUT_DF.info(buf=buffer, max_cols=colcount)
      filecontents = header + divider + datetimestamp + buffer.getvalue()
      tofile = os.path.join(OUTFOLDER, f"{filename}.txt")
      with open(tofile, 'w', encoding='utf-8') as f: f.write(filecontents)
      print(f"<{funcname}> ...done.")
   return None

# To turn column indexes into names. Will remove multi-indexing.
# Must assign the output to a dataframe e.g. df = colnames_from_index(df).
def colnames_from_index(INPUT_DF):
   cols = list(INPUT_DF)
   cols_new = []
   for item in cols:
      if type(item) == str:   # Columns that already have names will be strings. Use unchanged.
         cols_new.append(item)
      else:   # Columns that are indexed or multi-indexed will appear as tuples. Turn them into strings joined by underscores.
         cols_new.append('_'.join(str(i) for i in item))   # Convert each element of tuple to string before joining. Avoids error if an element is nan.

   # Write dataframe with new column names
   dfmod = INPUT_DF
   dfmod.columns = cols_new
   return dfmod

#%% PATHS AND VARIABLES

CURRENT_FOLDER = os.getcwd()
PARENT_FOLDER = os.path.dirname(CURRENT_FOLDER)
GRANDPARENT_FOLDER = os.path.dirname(PARENT_FOLDER)

# Folder for shared code with Liverpool
ETHIOPIA_CODE_FOLDER = CURRENT_FOLDER
ETHIOPIA_OUTPUT_FOLDER = os.path.join(PARENT_FOLDER ,'Program outputs')
ETHIOPIA_DATA_FOLDER = os.path.join(PARENT_FOLDER ,'Data')

# Folder managed by Murdoch University with disease-specific parameters and other updates
MURDOCH_BASE_FOLDER = os.path.join(CURRENT_FOLDER ,'Disease specific attribution')
MURDOCH_SCENARIO_FOLDER = os.path.join(MURDOCH_BASE_FOLDER ,'scenarios')
MURDOCH_OUTPUT_FOLDER = os.path.join(MURDOCH_BASE_FOLDER ,'output')

DASH_DATA_FOLDER = os.path.join(GRANDPARENT_FOLDER ,'Dash App' ,'data')

#%% COMBINE SCENARIO RESULT FILES
'''
This imports and combines CSV files that are output from the compartmental model.
'''
# This function reads all CSV files with the given prefix_suffix pattern in the given folder
# and merges them into a single dataframe. It also adds labels for species, production system,
# year, and region.
def combine_ahle_scenarios(
        input_folder
        ,input_file_prefix      # String
        ,input_file_suffixes    # List of strings. Each one combined with input_file_prefix uniquely identifies a file to be read.
        ,label_species          # String: add column 'species' with this label
        ,label_prodsys          # String: add column 'production_system' with this label
        ,label_year             # Integer: add column 'year' with this value
        ,label_region           # String: add column 'region' with this value
    ):
    funcname = inspect.currentframe().f_code.co_name

    dfcombined = pd.DataFrame()   # Initialize merged data
    for i ,suffix in enumerate(input_file_suffixes):
        # Read file if it exists
        try:
            df = pd.read_csv(os.path.join(input_folder ,f'{input_file_prefix}_{suffix}.csv'))

            # Add column suffixes
            if suffix.upper() == 'ALL_MORTALITY_ZERO':      # Recode for consistency
                suffix = 'MORTALITY_ZERO'
            df = df.add_suffix(f'_{suffix}')
            df = df.rename(columns={f'Item_{suffix}':'Item' ,f'Group_{suffix}':'Group'})

            # Add to merged data
            if i == 0:
                dfcombined = df.copy()
            else:
                dfcombined = pd.merge(left=dfcombined ,right=df, on=['Item' ,'Group'] ,how='outer')
        except FileNotFoundError:
            print(f'<{funcname}> File not found: ' ,os.path.join(input_folder ,f'{input_file_prefix}_{suffix}.csv'))
            print(f'<{funcname}> Moving to next file.')

    # Add label columns
    dfcombined['species'] = label_species
    dfcombined['production_system'] = label_prodsys
    dfcombined['year'] = label_year
    dfcombined['region'] = label_region

    # Reorder columns
    cols_first = ['region' ,'species' ,'production_system' ,'year']
    cols_other = [i for i in list(dfcombined) if i not in cols_first]
    dfcombined = dfcombined.reindex(columns=cols_first + cols_other)

    # Cleanup column names
    cleancolnames(dfcombined)

    return dfcombined

# =============================================================================
#### *ORIGINAL* Small ruminants
# =============================================================================
'''
These scenarios have only been produced for a single year (2021) and at the
national level.

Update August 2023: these have been replaced with results from Murdoch's updated
compartmental model code, which saves all results in a single file.
'''
# small_rum_suffixes=[
#     'Current'

#     ,'Ideal'
#     ,'ideal_AF'
#     ,'ideal_AM'
#     ,'ideal_JF'
#     ,'ideal_JM'
#     ,'ideal_NF'
#     ,'ideal_NM'

#     # Disease-specific scenarios
#     ,'PPR'
#     ,'Bruc'

#     ,'all_mortality_zero'
#     ,'mortality_zero_AF'
#     ,'mortality_zero_AM'
#     ,'mortality_zero_J'
#     ,'mortality_zero_N'

#     ,'all_mort_25_imp'
#     ,'mort_25_imp_AF'
#     ,'mort_25_imp_AM'
#     ,'mort_25_imp_J'
#     ,'mort_25_imp_N'

#     ,'all_mort_50_imp'
#     ,'mort_50_imp_AF'
#     ,'mort_50_imp_AM'
#     ,'mort_50_imp_J'
#     ,'mort_50_imp_N'

#     ,'all_mort_75_imp'
#     ,'mort_75_imp_AF'
#     ,'mort_75_imp_AM'
#     ,'mort_75_imp_J'
#     ,'mort_75_imp_N'

#     ,'Current_growth_25_imp_All'
#     ,'Current_growth_25_imp_AF'
#     ,'Current_growth_25_imp_AM'
#     ,'Current_growth_25_imp_JF'
#     ,'Current_growth_25_imp_JM'
#     ,'Current_growth_25_imp_NF'
#     ,'Current_growth_25_imp_NM'

#     ,'Current_growth_50_imp_All'
#     ,'Current_growth_50_imp_AF'
#     ,'Current_growth_50_imp_AM'
#     ,'Current_growth_50_imp_JF'
#     ,'Current_growth_50_imp_JM'
#     ,'Current_growth_50_imp_NF'
#     ,'Current_growth_50_imp_NM'

#     ,'Current_growth_75_imp_All'
#     ,'Current_growth_75_imp_AF'
#     ,'Current_growth_75_imp_AM'
#     ,'Current_growth_75_imp_JF'
#     ,'Current_growth_75_imp_JM'
#     ,'Current_growth_75_imp_NF'
#     ,'Current_growth_75_imp_NM'

#     ,'Current_growth_100_imp_All'
#     ,'Current_growth_100_imp_AF'
#     ,'Current_growth_100_imp_AM'
#     ,'Current_growth_100_imp_JF'
#     ,'Current_growth_100_imp_JM'
#     ,'Current_growth_100_imp_NF'
#     ,'Current_growth_100_imp_NM'

#     ,'Current_repro_25_imp'
#     ,'Current_repro_50_imp'
#     ,'Current_repro_75_imp'
#     ,'Current_repro_100_imp'
# ]

# ahle_sheep_clm = combine_ahle_scenarios(
#     input_folder=os.path.join(ETHIOPIA_OUTPUT_FOLDER ,'ahle SMALL RUMINANTS')
#     ,input_file_prefix='ahle_CLM_S'
#     ,input_file_suffixes=small_rum_suffixes
#     ,label_species='Sheep'
#     ,label_prodsys='Crop livestock mixed'
#     ,label_year=2021
#     ,label_region='National'
# )
# datainfo(ahle_sheep_clm)

# ahle_sheep_past = combine_ahle_scenarios(
#     input_folder=os.path.join(ETHIOPIA_OUTPUT_FOLDER ,'ahle SMALL RUMINANTS')
#     ,input_file_prefix='ahle_Past_S'
#     ,input_file_suffixes=small_rum_suffixes
#     ,label_species='Sheep'
#     ,label_prodsys='Pastoral'
#     ,label_year=2021
#     ,label_region='National'
# )
# datainfo(ahle_sheep_past)

# ahle_goat_clm = combine_ahle_scenarios(
#     input_folder=os.path.join(ETHIOPIA_OUTPUT_FOLDER ,'ahle SMALL RUMINANTS')
#     ,input_file_prefix='ahle_CLM_G'
#     ,input_file_suffixes=small_rum_suffixes
#     ,label_species='Goat'
#     ,label_prodsys='Crop livestock mixed'
#     ,label_year=2021
#     ,label_region='National'
# )
# datainfo(ahle_goat_clm)

# ahle_goat_past = combine_ahle_scenarios(
#     input_folder=os.path.join(ETHIOPIA_OUTPUT_FOLDER ,'ahle SMALL RUMINANTS')
#     ,input_file_prefix='ahle_Past_G'
#     ,input_file_suffixes=small_rum_suffixes
#     ,label_species='Goat'
#     ,label_prodsys='Pastoral'
#     ,label_year=2021
#     ,label_region='National'
# )
# datainfo(ahle_goat_past)

# # Adjust column names to match other species
# ahle_sheep_clm.columns = ahle_sheep_clm.columns.str.replace('_all_mortality_zero' ,'_mortality_zero')
# ahle_sheep_past.columns = ahle_sheep_past.columns.str.replace('_all_mortality_zero' ,'_mortality_zero')
# ahle_goat_clm.columns = ahle_goat_clm.columns.str.replace('_all_mortality_zero' ,'_mortality_zero')
# ahle_goat_past.columns = ahle_goat_past.columns.str.replace('_all_mortality_zero' ,'_mortality_zero')

# =============================================================================
#### Small ruminants
# =============================================================================
'''
Updated small ruminant results have all scenarios in a single file.
'''
# Import
ahle_sr_imp = pd.read_csv(os.path.join(MURDOCH_OUTPUT_FOLDER ,'ahle_sr.csv'))
cleancolnames(ahle_sr_imp)
datainfo(ahle_sr_imp)

# -----------------------------------------------------------------------------
# Modify for consistency with later data processing
# -----------------------------------------------------------------------------
# Pivot scenario labels into columns
# To match cattle and poultry files
ahle_sr_p = ahle_sr_imp.pivot(
	index=['species' ,'system' ,'item' ,'group']
	,columns=['scenario']
	,values=['mean' ,'stdev' ,'min' ,'q1' ,'median' ,'q3' ,'max']
)
ahle_sr_p = colnames_from_index(ahle_sr_p) 	# If multi-indexed columns were created, flatten index
ahle_sr_p = ahle_sr_p.reset_index()           # Pivoting will change columns to indexes. Change them back.
cleancolnames(ahle_sr_p)

# Rename columns
ahle_sr_p = ahle_sr_p.rename(columns={'system':'production_system'})
ahle_sr_p['year'] = 2021
ahle_sr_p['region'] = 'National'

# Recode species, production systems, and groups
# Note items already match
recode_sr_species = {
    "G":"Goat"
    ,"S":"Sheep"
}
ahle_sr_p['species'] = ahle_sr_p['species'].replace(recode_sr_species)

recode_sr_prodsys = {
    "CLM":"Crop livestock mixed"
    ,"Past":"Pastoral"
}
ahle_sr_p['production_system'] = ahle_sr_p['production_system'].replace(recode_sr_prodsys)

recode_sr_groups = {
    "Female Combined":"Overall Female"
    ,"Male Combined":"Overall Male"
}
ahle_sr_p['group'] = ahle_sr_p['group'].replace(recode_sr_groups)

datainfo(ahle_sr_p)

# =============================================================================
#### Cattle base
# =============================================================================
cattle_suffixes = [
    # August 2023: these are based on the most recent cattle parameters from Murdoch University
    'Current'
    ,'Ideal'
    ,'Bruc'
    ,'FMD'

    # August 2023: these are based on legacy scenario parameters
    ,'ideal_AF'
    ,'ideal_AM'
    ,'ideal_JF'
    ,'ideal_JM'
    ,'ideal_NF'
    ,'ideal_NM'
    ,'ideal_O'

    ,'mortality_zero'
    ,'mortality_zero_AF'
    ,'mortality_zero_AM'
    ,'mortality_zero_J'
    ,'mortality_zero_N'
    ,'mortality_zero_O'
]

# Import CLM
ahle_cattle_clm = combine_ahle_scenarios(
    input_folder=os.path.join(ETHIOPIA_OUTPUT_FOLDER ,'ahle CATTLE')
    ,input_file_prefix='ahle_CLM_C'
    ,input_file_suffixes=cattle_suffixes
    ,label_species='Cattle'
    ,label_prodsys='Crop livestock mixed'
    ,label_year=2021
    ,label_region='National'
    )
datainfo(ahle_cattle_clm ,130)

# Import pastoral
ahle_cattle_past = combine_ahle_scenarios(
    input_folder=os.path.join(ETHIOPIA_OUTPUT_FOLDER ,'ahle CATTLE')
    ,input_file_prefix='ahle_Past_C'
    ,input_file_suffixes=cattle_suffixes
    ,label_species='Cattle'
    ,label_prodsys='Pastoral'
    ,label_year=2021
    ,label_region='National'
    )
datainfo(ahle_cattle_past ,130)

# Import periurban dairy
ahle_cattle_peri = combine_ahle_scenarios(
    input_folder=os.path.join(ETHIOPIA_OUTPUT_FOLDER ,'ahle CATTLE')
    ,input_file_prefix='ahle_PUD_C'
    ,input_file_suffixes=cattle_suffixes
    ,label_species='Cattle'
    ,label_prodsys='Periurban dairy'
    ,label_year=2021
    ,label_region='National'
    )
datainfo(ahle_cattle_peri ,130)

# =============================================================================
#### Cattle Yearly
# =============================================================================
'''
August 2023: Yearly scenarios have not been updated, so may be inconsistent with
the single-year results.

This includes a loop to import each year and append to a master cattle dataframe.
'''
cattle_suffixes_yearlyandregional = [
    'current'

    ,'ideal'
    ,'ideal_AF'
    ,'ideal_AM'
    ,'ideal_JF'
    ,'ideal_JM'
    ,'ideal_NF'
    ,'ideal_NM'
    ,'ideal_O'

    ,'all_mortality_zero'
    ,'mortality_zero'
    ,'mortality_zero_AF'
    ,'mortality_zero_AM'
    ,'mortality_zero_J'
    ,'mortality_zero_N'
    ,'mortality_zero_O'

    # Disease-specific scenarios
    ,'Bruc'
]

ahle_cattle_yearly_aslist = []         # Initialize
for YEAR in range(2017 ,2022):
    # Import CLM
    ahle_cattle_clm_yearly = combine_ahle_scenarios(
        input_folder=os.path.join(ETHIOPIA_OUTPUT_FOLDER ,'ahle CATTLE' ,'Yearly results' ,f"{YEAR}")
        ,input_file_prefix='ahle_cattle_trial_CLM'
        ,input_file_suffixes=cattle_suffixes_yearlyandregional
        ,label_species='Cattle'
        ,label_prodsys='Crop livestock mixed'
        ,label_year=YEAR
        ,label_region='National'
        )
    datainfo(ahle_cattle_clm_yearly ,120)

 	# Turn into list and append to master
    ahle_cattle_clm_yearly_aslist = ahle_cattle_clm_yearly.to_dict(orient='records')
    ahle_cattle_yearly_aslist.extend(ahle_cattle_clm_yearly_aslist)
    del ahle_cattle_clm_yearly_aslist

    # Import pastoral
    ahle_cattle_past_yearly = combine_ahle_scenarios(
        input_folder=os.path.join(ETHIOPIA_OUTPUT_FOLDER ,'ahle CATTLE' ,'Yearly results' ,f"{YEAR}")
        ,input_file_prefix='ahle_cattle_trial_past'
        ,input_file_suffixes=cattle_suffixes_yearlyandregional
        ,label_species='Cattle'
        ,label_prodsys='Pastoral'
        ,label_year=YEAR
        ,label_region='National'
        )
    datainfo(ahle_cattle_past_yearly ,120)

 	# Turn into list and append to master
    ahle_cattle_past_yearly_aslist = ahle_cattle_past_yearly.to_dict(orient='records')
    ahle_cattle_yearly_aslist.extend(ahle_cattle_past_yearly_aslist)
    del ahle_cattle_past_yearly_aslist

    # Import periurban dairy
    ahle_cattle_peri_yearly = combine_ahle_scenarios(
        input_folder=os.path.join(ETHIOPIA_OUTPUT_FOLDER ,'ahle CATTLE' ,'Yearly results' ,f"{YEAR}")
        ,input_file_prefix='ahle_cattle_trial_periurban_dairy'
        ,input_file_suffixes=cattle_suffixes_yearlyandregional
        ,label_species='Cattle'
        ,label_prodsys='Periurban dairy'
        ,label_year=YEAR
        ,label_region='National'
        )
    datainfo(ahle_cattle_peri_yearly ,120)

 	# Turn into list and append to master
    ahle_cattle_peri_yearly_aslist = ahle_cattle_peri_yearly.to_dict(orient='records')
    ahle_cattle_yearly_aslist.extend(ahle_cattle_peri_yearly_aslist)
    del ahle_cattle_peri_yearly_aslist

# Convert master list into data frame
ahle_cattle_yearly = pd.DataFrame.from_dict(ahle_cattle_yearly_aslist ,orient='columns')
del ahle_cattle_yearly_aslist
datainfo(ahle_cattle_yearly ,120)

# =============================================================================
#### Cattle Regional
# =============================================================================
'''
August 2023: Regional scenarios have not been updated, but I am keeping these
in the data for the regional map in Dash.

These scenarios have been run for regions within Ethiopia, so this uses a loop
to import each region and append to a master dataframe.

These have not been run for different years. They are being assigned year 2021.
'''
# Should match list defined in 1_run_ahle_simulation_standalone.py
list_eth_regions = [
    'Afar'
    ,'Amhara'
    ,'BG'
    ,'Gambella'
    ,'Oromia'
    ,'Sidama'
    ,'SNNP'
    ,'Somali'
    ,'Tigray'
    ]

ahle_cattle_regional_aslist = []        # Initialize
for REGION in list_eth_regions:
    # Import CLM
    ahle_cattle_regional_clm = combine_ahle_scenarios(
        input_folder=os.path.join(ETHIOPIA_OUTPUT_FOLDER ,'ahle CATTLE' ,'Subnational results' ,f'{REGION}')
        ,input_file_prefix='ahle_cattle_trial_CLM'
        ,input_file_suffixes=cattle_suffixes_yearlyandregional
        ,label_species='Cattle'
        ,label_prodsys='Crop livestock mixed'
        ,label_year=2021
        ,label_region=f'{REGION}'
        )
    datainfo(ahle_cattle_regional_clm ,120)

	# Turn into list and append to master
    ahle_cattle_regional_clm_aslist = ahle_cattle_regional_clm.to_dict(orient='records')
    ahle_cattle_regional_aslist.extend(ahle_cattle_regional_clm_aslist)
    del ahle_cattle_regional_clm_aslist

    # Import pastoral
    ahle_cattle_regional_past = combine_ahle_scenarios(
        input_folder=os.path.join(ETHIOPIA_OUTPUT_FOLDER ,'ahle CATTLE' ,'Subnational results' ,f'{REGION}')
        ,input_file_prefix='ahle_cattle_trial_past'
        ,input_file_suffixes=cattle_suffixes_yearlyandregional
        ,label_species='Cattle'
        ,label_prodsys='Pastoral'
        ,label_year=2021
        ,label_region=f'{REGION}'
        )
    datainfo(ahle_cattle_regional_past ,120)

	# Turn into list and append to master
    ahle_cattle_regional_past_aslist = ahle_cattle_regional_past.to_dict(orient='records')
    ahle_cattle_regional_aslist.extend(ahle_cattle_regional_past_aslist)
    del ahle_cattle_regional_past_aslist

    # Import periurban dairy
    ahle_cattle_regional_peri = combine_ahle_scenarios(
        input_folder=os.path.join(ETHIOPIA_OUTPUT_FOLDER ,'ahle CATTLE' ,'Subnational results' ,f'{REGION}')
        ,input_file_prefix='ahle_cattle_trial_periurban_dairy'
        ,input_file_suffixes=cattle_suffixes_yearlyandregional
        ,label_species='Cattle'
        ,label_prodsys='Periurban dairy'
        ,label_year=2021
        ,label_region=f'{REGION}'
        )
    datainfo(ahle_cattle_regional_peri ,120)

	# Turn into list and append to master
    ahle_cattle_regional_peri_aslist = ahle_cattle_regional_peri.to_dict(orient='records')
    ahle_cattle_regional_aslist.extend(ahle_cattle_regional_peri_aslist)
    del ahle_cattle_regional_peri_aslist

# Convert master list into data frame
ahle_cattle_regional = pd.DataFrame.from_dict(ahle_cattle_regional_aslist ,orient='columns')
del ahle_cattle_regional_aslist
datainfo(ahle_cattle_regional ,120)

# Recode region names to match those in geojson for mapping
rename_regions = {
    'Afar':'Afar'
    ,'Amhara':'Amhara'
    ,'BG':'Benishangul Gumz'
    ,'Gambella':'Gambela'
    ,'Oromia':'Oromia'
    ,'Sidama':'Sidama'
    ,'SNNP':'SNNP'
    ,'Somali':'Somali'
    ,'Tigray':'Tigray'
    }
ahle_cattle_regional['region'] = ahle_cattle_regional['region'].replace(rename_regions)

# Create values for South West Ethiopia by replicating SNNP (they were the same region until recently)
ahle_cattle_region_swe = ahle_cattle_regional.query("region == 'SNNP'").copy()
ahle_cattle_region_swe['region'] = 'South West Ethiopia'
ahle_cattle_regional = pd.concat([ahle_cattle_regional ,ahle_cattle_region_swe] ,ignore_index=True)

# =============================================================================
#### Poultry
# =============================================================================
'''
These scenarios have only been produced for a single year (2021) and at the
national level.
'''
poultry_suffixes = [
    'current'
    ,'ideal'
    ,'ideal_A'
    ,'ideal_J'
    ,'ideal_N'
    ,'mortality_zero'
    ,'mortality_zero_A'
    ,'mortality_zero_J'
    ,'mortality_zero_N'
]

ahle_poultry_smallholder = combine_ahle_scenarios(
    input_folder=os.path.join(ETHIOPIA_OUTPUT_FOLDER ,'ahle POULTRY')
    ,input_file_prefix='ahle_Smallholder_hybrid'
    ,input_file_suffixes=poultry_suffixes
    ,label_species='Poultry hybrid'
    ,label_prodsys='Small holder'
    ,label_year=2021
    ,label_region='National'
)
datainfo(ahle_poultry_smallholder)

ahle_poultry_villagehybrid = combine_ahle_scenarios(
    input_folder=os.path.join(ETHIOPIA_OUTPUT_FOLDER ,'ahle POULTRY')
    ,input_file_prefix='ahle_Village_hybrid'
    ,input_file_suffixes=poultry_suffixes
    ,label_species='Poultry hybrid'
    ,label_prodsys='Village'
    ,label_year=2021
    ,label_region='National'
)
datainfo(ahle_poultry_villagehybrid)

ahle_poultry_villageindig = combine_ahle_scenarios(
    input_folder=os.path.join(ETHIOPIA_OUTPUT_FOLDER ,'ahle POULTRY')
    ,input_file_prefix='ahle_Village_indigenous'
    ,input_file_suffixes=poultry_suffixes
    ,label_species='Poultry indigenous'
    ,label_prodsys='Village'
    ,label_year=2021
    ,label_region='National'
)
datainfo(ahle_poultry_villageindig)

# =============================================================================
#### Stack all
# =============================================================================
concat_list = [
    # *ORIGINAL Small ruminants*
    # ahle_sheep_clm
    # ,ahle_sheep_past
    # ,ahle_goat_clm
    # ,ahle_goat_past

    # Small ruminants
    ahle_sr_p

    # Cattle base
    ,ahle_cattle_clm
    ,ahle_cattle_past
    ,ahle_cattle_peri

    # Cattle Yearly
    ,ahle_cattle_yearly

    # Cattle Regional
    ,ahle_cattle_regional

    # Poultry
    ,ahle_poultry_smallholder
    ,ahle_poultry_villagehybrid
    ,ahle_poultry_villageindig
]
ahle_combo_cat = pd.concat(
   concat_list      # List of dataframes to concatenate
   ,axis=0              # axis=0: concatenate rows (stack), axis=1: concatenate columns (merge)
   ,join='outer'        # 'outer': keep all index values from all data frames
   ,ignore_index=True   # True: do not keep index values on concatenation axis
)

# Reset dataframe with a copy due to warning about fragmented data
ahle_combo = ahle_combo_cat.copy()
del ahle_combo_cat

# Split age and sex groups into their own columns
ahle_combo[['age_group' ,'sex']] = ahle_combo['group'].str.split(' ' ,expand=True)

# Recode sex
recode_sex = {
    'Combined':'Overall'
    ,np.nan:'Overall'
    }
ahle_combo['sex'] = ahle_combo['sex'].replace(recode_sex)

# Special handling for Oxen
ahle_combo.loc[ahle_combo['group'].str.upper() == 'OXEN' ,'age_group'] = 'Oxen'
ahle_combo.loc[ahle_combo['group'].str.upper() == 'OXEN' ,'sex'] = 'Male'

# Reorder columns
cols_first = ['species' ,'region' ,'production_system' ,'item' ,'group' ,'age_group' ,'sex' ,'year']
cols_other = [i for i in list(ahle_combo) if i not in cols_first]
ahle_combo = ahle_combo.reindex(columns=cols_first + cols_other)

# =============================================================================
#### Export
# =============================================================================
datainfo(ahle_combo ,200)

ahle_combo.to_csv(os.path.join(ETHIOPIA_OUTPUT_FOLDER ,'ahle_all_stacked.csv') ,index=False)

#%% CHECKS ON RAW SIMULATION OUTPUT
'''
This creates some check_ datasets and prints checks to the console.
'''
check_ahle_combo = ahle_combo.copy()

_group_overall = (check_ahle_combo['group'].str.upper() == 'OVERALL')
_item_grossmargin = (check_ahle_combo['item'].str.upper() == 'GROSS MARGIN')
summarize_byvars = ('species' ,'region' ,'production_system' ,'year')

# =============================================================================
#### Change in Gross Margin
# =============================================================================
check_grossmargin_overall = check_ahle_combo.loc[_group_overall].loc[_item_grossmargin]
check_grossmargin_overall = check_grossmargin_overall.eval(
    # Change in Gross Margin overall vs. individual ideal scenarios
    '''
    gmchange_ideal_overall = mean_ideal - mean_current

    gmchange_ideal_af = mean_ideal_af - mean_current
    gmchange_ideal_am = mean_ideal_am - mean_current
    gmchange_ideal_jf = mean_ideal_jf - mean_current
    gmchange_ideal_jm = mean_ideal_jm - mean_current
    gmchange_ideal_nf = mean_ideal_nf - mean_current
    gmchange_ideal_nm = mean_ideal_nm - mean_current
    gmchange_ideal_o = mean_ideal_o - mean_current
    '''
    # Mortality as proportion of total AHLE
    '''
    gmchange_dueto_mortality = mean_mortality_zero - mean_current
    gmchange_dueto_production = gmchange_ideal_overall - gmchange_dueto_mortality
    gmchange_dueto_mortality_prpn = gmchange_dueto_mortality / gmchange_ideal_overall
    '''
    # Change in gross margin due to specific diseases
    '''
    gmchange_dueto_ppr = mean_ideal - mean_ppr
    gmchange_dueto_ppr_ratio = gmchange_dueto_ppr / gmchange_ideal_overall

    gmchange_dueto_bruc = mean_ideal - mean_bruc
    gmchange_dueto_bruc_ratio = gmchange_dueto_bruc / gmchange_ideal_overall

    gmchange_dueto_fmd = mean_ideal - mean_fmd
    gmchange_dueto_fmd_ratio = gmchange_dueto_fmd / gmchange_ideal_overall
    '''
)
# Not all agesex scenarios apply to every species. Set missing to zero.
gmchange_ind_list = [
    'gmchange_ideal_af'
    ,'gmchange_ideal_am'
    ,'gmchange_ideal_jf'
    ,'gmchange_ideal_jm'
    ,'gmchange_ideal_nf'
    ,'gmchange_ideal_nm'
    ,'gmchange_ideal_o'
]
for COL in gmchange_ind_list:
    check_grossmargin_overall[COL] = check_grossmargin_overall[COL].fillna(0)

check_grossmargin_overall = check_grossmargin_overall.eval(
    '''
    gmchange_ideal_sumind = gmchange_ideal_af + gmchange_ideal_am \
        + gmchange_ideal_jf + gmchange_ideal_jm \
            + gmchange_ideal_nf + gmchange_ideal_nm \
                + gmchange_ideal_o
    gmchange_ideal_check = gmchange_ideal_sumind / gmchange_ideal_overall
    '''
)

print('\n> Summarizing the change in Gross Margin for ideal overall')
print(check_grossmargin_overall['gmchange_ideal_overall'].describe())
print(check_grossmargin_overall[list(summarize_byvars) + ['gmchange_ideal_overall']])

print('\n> Checking the change in Gross Margin for ideal overall vs. individual ideal scenarios')
print(check_grossmargin_overall[list(summarize_byvars) + ['gmchange_ideal_check']])

print('\n> Checking mortality as proportion of total AHLE')
print(check_grossmargin_overall[list(summarize_byvars) + ['gmchange_dueto_mortality_prpn']])

print('\n> Checking the change in gross margin due to PPR as proportion of total AHLE')
print(check_grossmargin_overall[list(summarize_byvars) + ['gmchange_dueto_ppr_ratio']])

print('\n> Checking the change in gross margin due to Brucellosis as proportion of total AHLE')
print(check_grossmargin_overall[list(summarize_byvars) + ['gmchange_dueto_bruc_ratio']])

print('\n> Checking the change in gross margin due to FMD as proportion of total AHLE')
print(check_grossmargin_overall[list(summarize_byvars) + ['gmchange_dueto_fmd_ratio']])

# =============================================================================
#### Sum of agesex groups compared to system total for each item
# =============================================================================
# Sum individual agesex groups for each item
_indiv_sex = (check_ahle_combo['group'].str.upper().isin(['ADULT FEMALE' ,'ADULT MALE' ,'JUVENILE FEMALE' ,'JUVENILE MALE' ,'NEONATAL FEMALE' ,'NEONATAL MALE' ,'OXEN']))
check_agesex_sums = pd.DataFrame(check_ahle_combo.loc[_indiv_sex]\
    .groupby(list(summarize_byvars) + ['item'] ,observed=True)['mean_current'].sum())
check_agesex_sums.columns = ['mean_current_sumagesex']

# Merge group total for each item
check_agesex_sums = pd.merge(
    left=check_agesex_sums
    ,right=check_ahle_combo.loc[_group_overall ,list(summarize_byvars) + ['item' ,'mean_current']]
    ,on=list(summarize_byvars) + ['item']
    ,how='left'
)
check_agesex_sums = check_agesex_sums.rename(columns={'mean_current':'mean_current_overall'})

# Calculate ratio
check_agesex_sums = check_agesex_sums.eval(
    '''
    check_ratio = mean_current_sumagesex / mean_current_overall
    '''
)
print('\n> Checking the sum of individual age/sex groups compared to the overall for each item')
print('\nMaximum ratio \n-------------')
print(check_agesex_sums.groupby(list(summarize_byvars))['check_ratio'].max())
print('\nMinimum ratio \n-------------')
print(check_agesex_sums.groupby(list(summarize_byvars))['check_ratio'].min())

# =============================================================================
#### Distributions
# =============================================================================
check_ahle_combo['expected_min_current'] = sps.norm.ppf(.01 ,loc=check_ahle_combo['mean_current'] ,scale=check_ahle_combo['stdev_current'])
check_ahle_combo['expected_q1_current'] = sps.norm.ppf(.25 ,loc=check_ahle_combo['mean_current'] ,scale=check_ahle_combo['stdev_current'])
check_ahle_combo['expected_median_current'] = sps.norm.ppf(.50 ,loc=check_ahle_combo['mean_current'] ,scale=check_ahle_combo['stdev_current'])
check_ahle_combo['expected_q3_current'] = sps.norm.ppf(.75 ,loc=check_ahle_combo['mean_current'] ,scale=check_ahle_combo['stdev_current'])
check_ahle_combo['expected_max_current'] = sps.norm.ppf(.99 ,loc=check_ahle_combo['mean_current'] ,scale=check_ahle_combo['stdev_current'])

datainfo(check_ahle_combo ,200)

#%% ADJUSTMENTS & PLACEHOLDERS

ahle_combo_adj = ahle_combo.copy()

# Define full set of key variables
# Make tuple so immutable
all_byvars = ('species' ,'region' ,'production_system' ,'item' ,'item_type_code' ,'group' ,'age_group' ,'sex' ,'year')

# =============================================================================
#### Add item type code
# =============================================================================
# pq = physical quantity, mv = monetary value, mc = monetary cost
item_type_code = {
    'Num Offtake':'pq'
    ,'Cml Pop Growth':'pq'
    ,'Total Number Increase':'pq'
    ,'Total Mortality':'pq'
    ,'Population Liveweight (kg)':'pq'
    ,'Offtake Liveweight (kg)':'pq'
    ,'Meat (kg)':'pq'
    ,'Manure':'pq'
    ,'Hides':'pq'
    ,'Milk':'pq'
    ,'Wool':'pq'
    ,'Cml Dry Matter':'pq'
    ,'Eggs Sold':'pq'
    ,'Eggs Consumed':'pq'

    ,'Births':'pq'
    ,'Culls':'pq'
    ,'Deaths':'pq'
    ,'Difference':'pq'
    ,'Growth in':'pq'
    ,'Growth out':'pq'
    ,'Inflow':'pq'
    ,'Offtakes':'pq'
    ,'Outflow':'pq'
    ,'Population':'pq'
    ,'Sum':'pq'

    ,'Value of Offtake':'mv'
    ,'Value of Herd Increase':'mv'
    ,'Value of Herd Increase plus Offtake':'mv'
    ,'Value of Manure':'mv'
    ,'Value of Hides':'mv'
    ,'Value of Milk':'mv'
    ,'Total Production Value':'mv'
    ,'Gross Margin':'mv'
    ,'Value of draught':'mv'
    ,'Value of Eggs sold':'mv'
    ,'Value of Eggs consumed':'mv'

    ,'Feed Cost':'mc'
    ,'Labour Cost':'mc'
    ,'Health Cost':'mc'
    ,'Capital Cost':'mc'
    ,'Value of Total Mortality':'mc'
    ,'Infrastructure Cost':'mc'
    ,'Total Expenditure':'mc'
}
ahle_combo_adj['item_type_code'] = ahle_combo_adj['item'].replace(item_type_code)

# =============================================================================
#### Make all monetary cost items negative
# =============================================================================
float_cols = list(ahle_combo_adj.select_dtypes(include='float'))
for COL in float_cols:
    ahle_combo_adj[COL] = np.where(
        ahle_combo_adj['item_type_code'] == 'mc'
        ,ahle_combo_adj[COL] * -1
        ,ahle_combo_adj[COL]
    )

# Reorder columns
cols_first = list(all_byvars)
cols_other = [i for i in list(ahle_combo_adj) if i not in cols_first]
ahle_combo_adj = ahle_combo_adj.reindex(columns=cols_first + cols_other)

# =============================================================================
#### Add yearly placeholder rows
# =============================================================================
'''
Goal: add yearly placeholder values for any species, production system, item,
and group that does not have them. Keep actual yearly values if they exist.
'''
# Each numeric column gets inflated/deflated by a percentage
yearly_adjustment = 1.05    # Desired yearly change in values

# Get list of columns for which to add placeholders
vary_by_year = list(ahle_combo_adj.select_dtypes(include='float'))  # All columns of type Float

# Turn data into list
# ahle_combo_adj_plhdyear = ahle_combo_adj.loc[ahle_combo_adj['region'] == 'National']    # Only creating yearly placeholders for national results, not regional
ahle_combo_adj_plhdyear = ahle_combo_adj.copy()
ahle_combo_adj_plhdyear_aslist = ahle_combo_adj_plhdyear.to_dict(orient='records')

base_year = 2021
create_years = list(range(2017 ,2022))
for YEAR in create_years:
    # Create dataset for this year
    single_year_df = ahle_combo_adj_plhdyear.copy()
    single_year_df['year'] = YEAR

    # Adjust numeric columns
    adj_factor = yearly_adjustment**(YEAR - base_year)
    for COL in vary_by_year:
        single_year_df[COL] = single_year_df[COL] * adj_factor

    # Turn data into list and append
    single_year_df_aslist = single_year_df.to_dict(orient='records')
    ahle_combo_adj_plhdyear_aslist.extend(single_year_df_aslist)

# Convert list of dictionaries into data frame
ahle_combo_adj_plhdyear = pd.DataFrame.from_dict(ahle_combo_adj_plhdyear_aslist ,orient='columns')
del ahle_combo_adj_plhdyear_aslist ,single_year_df ,single_year_df_aslist

# Concatenate with original
ahle_combo_adj = pd.concat([ahle_combo_adj ,ahle_combo_adj_plhdyear] ,axis=0 ,ignore_index=True)
del ahle_combo_adj_plhdyear

# Remove duplicate values, keeping the first (the first is the actual value for that year if it exists)
ahle_combo_adj = ahle_combo_adj.drop_duplicates(
    subset=all_byvars
    ,keep='first'
).reset_index(drop=True)

# =============================================================================
#### Export
# =============================================================================
datainfo(ahle_combo_adj ,200)

ahle_combo_adj.to_csv(os.path.join(ETHIOPIA_OUTPUT_FOLDER ,'ahle_all_stacked_adj.csv') ,index=False)
